requirements.txt
django>=3.0
pillow
django-crispy-forms
boto3
gunicorn
django-markdownify
django-tinymce
django-froala-editor
django-storages
django-ses
python-slugify
django-polymorphic
django-recaptcha
qrcode
django-ckeditor
#django-nested-inline
django-nested-admin
python-slugify
